+++ CHANGES.md
@@ -7,7 +7,8 @@ but this file may sometimes contain later improvements (e.g. typo fixes).
 
 ## next (not yet released)
 
-No changes yet.
+- Added the `responseBoolean` utility function,
+  to get a boolean out of a response object regardless of `formatversion`.
 
 ## v0.3.0 (2021-10-10)